Improve resolution + fix crash
Crash:
```c++
_voltage = (voltage_pulse_width > 0) ? _voltage_multiplier / voltage_pulse_width / 2 : 0;
```
Since `voltage_pulse_width` is volatile, it may be set to 0 in an interrupt handling right after it was checked for not being zero.


Resolution improvement:
When measuring 230V mains voltage on a Sonoff POW r1, it reported the voltage with a resolution of about 8V.
Changing some types to float instead of int does improve it slightly.

The actual measurement was done by measuring the "last" pulse of a measurement period due to instability of the pulse frequency right after setting it.
For 222 & 230V mains this was a pulse width of roughly 970 - 1000 usec. This is also roughly the jitter of an ESP8266 when handling interrupts.

The change I made was to keep track of both the measurement duration, the last pulse width and count the nr. of pulses.
If the nr of pulses is low, the pulse duration is way longer than the jitter, so then we can take the last pulse width.
If the nr. of pulses is high, then the outliers of the first few pulses won't matter a lot. Also "instability" does work both ways and thus evens out.

Now the resolution of the measurements from the CF1 pin is way more usable and also quite accurate.
Compared it with the CSE7766 used in a Sonoff POW r2 and calibrated both using a proper multimeter.
Both differ in the order of 0.1V from each other and the multimeter.